modified:   Assets/Scenes/Main.unity 	modified:   Assets/Scenes/Menu.unity 	modified:   Assets/Scripts/GameMode Scripts/GM_Base.cs 	modified:   Assets/Scripts/GameMode Scripts/GM_Manager.cs 	modified:   Assets/Scripts/Hexgrid/HexGrid.cs 	modified:   Assets/Scripts/Network scripts/Network_DisplayScore.cs 	modified:   Assets/Scripts/Network scripts/Network_Manager.cs 	modified:   Assets/Scripts/Player scripts/Player_Score.cs 	Assets/Resources/Images/Thumbs.db 	Assets/Resources/Images/Thumbs.db.meta 	Assets/Resources/Images/koth_colormap.png 	Assets/Resources/Images/koth_colormap.png.meta 	Assets/Resources/Images/koth_heightmap.png 	Assets/Resources/Images/koth_heightmap.png.meta 	Assets/Resources/Materials/Indicator.mat 	Assets/Resources/Materials/Indicator.mat.meta 	Assets/Resources/Models/DFS Arena (1).ma 	Assets/Resources/Models/DFS Arena (1).ma.meta 	Assets/Resources/Models/DFS Arena.ma 	Assets/Resources/Models/DFS Arena.ma.meta 	Assets/Resources/Models/Materials/TestZone Symbolen.mat 	Assets/Resources/Models/Materials/TestZone Symbolen.mat.meta 	Assets/Resources/Models/Materials/lambert10.mat 	Assets/Resources/Models/Materials/lambert10.mat.meta 	Assets/Resources/Models/Materials/lambert6.mat 	Assets/Resources/Models/Materials/lambert6.mat.meta 	Assets/Resources/Models/Materials/lambert9.mat 	Assets/Resources/Models/Materials/lambert9.mat.meta 	Assets/Resources/Particles/PulsegunIn.prefab 	Assets/Resources/Particles/PulsegunIn.prefab.meta 	Assets/Resources/Particles/PulsegunOut.prefab 	Assets/Resources/Particles/PulsegunOut.prefab.meta 	Assets/Resources/Particles/arena.meta 	Assets/Resources/Particles/arena/ 	Assets/Resources/Prefabs/IndicatorSphere.prefab 	Assets/Resources/Prefabs/IndicatorSphere.prefab.meta 	Assets/Scripts/GameMode Scripts/Bases.meta 	Assets/Scripts/GameMode Scripts/Bases/

+++ Assets/Scripts/Network scripts/Network_DisplayScore.cs
@@ -20,7 +20,7 @@ public class Network_DisplayScore : NetworkBehaviour
                     if (pName != "")
                         pName = pName.Remove(pName.Length - 1);
 
-                    scoreText.GetComponent<Text>().text = pName + ": " + players [p].GetComponent<Player_Score>().Score;
+                    scoreText.GetComponent<Text>().text = pName + ": " + (int)players [p].GetComponent<Player_Score>().Score;
                 } else
                 {
                     scoreText.GetComponent<Text>().text = "";
